updates mmdet backbones
