+++ icevision/models/mmdet/models/retinanet/backbones/resnet_fpn.py
@@ -15,18 +15,6 @@ from icevision.imports import *
 from icevision.models.mmdet.utils import *
 
 
-class MMDetBackboneConfig:
-    def __init__(self, model_name, config_path, weights_url):
-        self.model_name = model_name
-        self.config_path = config_path
-        self.weights_url = weights_url
-        self.pretrained: bool
-
-    def __call__(self, pretrained):
-        self.pretrained = pretrained
-        return self
-
-
 class MMDetRetinanetBackboneConfig(MMDetBackboneConfig):
     def __init__(self, **kwargs):
         super().__init__(model_name="retinanet", **kwargs)
